Make backup folder for nano(1) in .kshrc
You can't just test to see if the binary is there, because everything
ships nano.  Just make the damn thing so it doesn't freak out if you
happen to run it.

In typical GNU-fashion, it's not smart enough to create it for itself
if it's not there and it's been told to use it.