Added a more-better-working dirstack implementation
taken from https://stackoverflow.com/questions/984635/pushd-popd-on-ksh
specifically https://stackoverflow.com/a/57021646/13100156

cleaned it up so that it could deal with spaces in directory names and
so that it didn't quite sprinkle so many variables around to conflict
with things.  needs more work to make as many variables as possible local.